Separate D3 utils to placate jest
