+++ js/job.jsx
@@ -17,9 +17,11 @@ import {
   FAILED_STATES,
   jobState,
   lolhadoop,
+} from './utils/utils';
+import {
   secondFormat,
   timeFormat,
-} from './utils';
+} from './utils/d3';
 import {notAvailable} from './mr';
 
 const {$, _, d3} = window;
