+++ package.json
@@ -31,12 +31,10 @@
     "vinyl-source-stream": "^2.0.0"
   },
   "dependencies": {
-    "d3": "^3.4.13",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-router": "^3.0.2",
-    "react-tooltip": "^3.4.0",
-    "underscore": "^1.6.0"
+    "react-tooltip": "^3.4.0"
   },
   "main": "gulpfile.js",
   "directories": {
